ml/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
torch==2.1.0
torchvision==0.16.0
Pillow==10.2.0
python-multipart==0.0.9
python-dotenv==1.0.1
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1
